Move remaining preprocess modules and test to commongrid; update all references to preprocess
