+++ echopype/__init__.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import, division, print_function
 
 from _echopype_version import version as __version__  # noqa
 
-from . import calibrate, consolidate, filter, mask, preprocess, utils
+from . import calibrate, consolidate, filter, mask, commongrid, utils
 from .convert.api import open_raw
 from .echodata.api import open_converted
 from .echodata.combine import combine_echodata
@@ -18,11 +18,11 @@ __all__ = [
     "open_raw",
     "open_converted",
     "combine_echodata",
+    "verbose",
     "calibrate",
+    "commongrid",
     "consolidate",
     "filter",
     "mask",
-    "preprocess",
     "utils",
-    "verbose",
 ]
